+++ tests/testthat/test-utils.R
@@ -126,7 +126,7 @@ test_that("cmdstan_summary works if bin/stansummary deleted file", {
     file.remove(file.path(cmdstan_path(), "bin", cmdstan_ext("stansummary")))
     fit_mcmc$cmdstan_summary()
   }
-  expect_output(delete_and_run(), "Inference for Stan model: logistic_model\\n2 chains: each with iter")
+  expect_output(delete_and_run(), "Inference for Stan model: logistic_model")
 })
 
 test_that("cmdstan_diagnose works if bin/diagnose deleted file", {
@@ -168,21 +168,21 @@ test_that("cmdstan_make_local() works", {
   expect_equal(cmdstan_make_local(), NULL)
   cpp_options = list(
    "CXX" = "clang++",
-   "CXXFLAGS+= -march-native",
+   "CXXFLAGS+= -march=native",
    TEST1 = TRUE,
    "TEST2" = FALSE
   )
   expect_equal(cmdstan_make_local(cpp_options = cpp_options),
                c(
                  "CXX=clang++",
-                 "CXXFLAGS+= -march-native",
+                 "CXXFLAGS+= -march=native",
                  "TEST1=true",
                  "TEST2=false"
                  ))
   expect_equal(cmdstan_make_local(cpp_options = list("TEST3" = TRUE)),
                c(
                  "CXX=clang++",
-                 "CXXFLAGS+= -march-native",
+                 "CXXFLAGS+= -march=native",
                  "TEST1=true",
                  "TEST2=false",
                  "TEST3=true"